Revamp qimages and instr_files to allow urls Rework item import/export to always use URLs for files

+++ course/managelibs.php
@@ -866,6 +866,7 @@ require("../footer.php");
 
 function delqimgs($qsid) {
   global $DBH;
+
   $srch_stm = $DBH->prepare("SELECT id FROM imas_qimages WHERE filename=:filename");
   $del_stm = $DBH->prepare("DELETE FROM imas_qimages WHERE id=:id");
 
@@ -878,9 +879,11 @@ function delqimgs($qsid) {
 		//DB $query = "SELECT id FROM imas_qimages WHERE filename='{$row[1]}'";
 		//DB $r2 = mysql_query($query) or die("Query failed :$query " . mysql_error());
 		//DB if (mysql_num_rows($r2)==1) {
-		$srch_stm->execute(array(':filename'=>$row[1]));
-		if ($srch_stm->rowCount()==1) {
-			unlink(rtrim(dirname(__FILE__), '/\\') .'/../assessment/qimages/'.$row[1]);
+		if (substr($row[1],0,4)!='http') {
+			$srch_stm->execute(array(':filename'=>$row[1]));
+			if ($srch_stm->rowCount()==1) {
+				unlink(rtrim(dirname(__FILE__), '/\\') .'/../assessment/qimages/'.$row[1]);
+			}
 		}
 		//DB $query = "DELETE FROM imas_qimages WHERE id='{$row[0]}'";
 		//DB mysql_query($query) or die("Query failed :$query " . mysql_error());
